+++ src/components/CompletedTodo.js
@@ -1,17 +1,19 @@
 import React from "react"
 
-const CompletedTodo = ({ todo, deleteTodo, completeTodo }) => {
-	return (
-		<li>
-			<input 
-				type='checkbox'
-				checked={todo.completed}
-				onChange={() => completeTodo(todo.id)}
-			/>
-			<del>{todo.title}</del>
-			<button type='button' onClick={() => deleteTodo(todo.id)}>Delete</button>
-		</li>
-	)
+class CompletedTodo extends React.Component {
+	render() {
+		return (
+			<li>
+				<input 
+					type='checkbox'
+					checked={this.props.todo.completed}
+					onChange={() => this.props.completeTodo(this.props.todo.id)}
+				/>
+				<del>{this.props.todo.title}</del>
+				<button type='button' onClick={() => this.props.deleteTodo(this.props.todo.id)}>Delete</button>
+			</li>
+		)
+	}
 }
 
 export default CompletedTodo
