+++ src/components/CompletedTodoList.js
@@ -1,22 +1,16 @@
 import React from "react"
-import CompletedTodo from "./CompletedTodo"
 
-const CompletedTodoList = ({ todos, deleteTodo, completeTodo }) => {
-	function sortModifiedAt(a, b) {
-		return b.modifiedAt - a.modifiedAt
+class CompletedTodoList extends React.Component {
+	render() {
+		return (
+			<div>
+				<h2>Completed ({this.props.todos.length})</h2>
+				<ul>
+					{this.props.children}
+				</ul>
+			</div>
+		)
 	}
-	return (
-		<ul>
-			{todos.sort(sortModifiedAt).map(todo =>
-				<CompletedTodo
-					key={todo.id}
-					todo={todo}
-					deleteTodo={deleteTodo}
-					completeTodo={completeTodo}
-				/>
-			)}
-		</ul>
-	)
 }
 
 export default CompletedTodoList
